backend/requirements.txt
fastapi
uvicorn[standard]
groq
motor
pydantic-settings
python-dotenv
pandas
numpy
scikit-learn
matplotlib
motor
pymongo
certifi
python-multipart
pypdf
sentence-transformers
langchain
langchain-mongodb
langchain-huggingface
